shift convert task function to utils.py
